Keep projected paths bound to shell style
